Handling duplicates and out of order delivery, test cases added

+++ cs733/assignment2/response.go
@@ -12,7 +12,7 @@ type Commit struct {
 }
 
 type Alarm struct {
-	t int64
+	t int
 }
 
 type LogStore struct {
